pipeline/pyproject.toml
[project]
name = "boston-pipeline"
version = "1.0.0"
description = "Daily data pipeline for Boston Urban Hazard Maps"
requires-python = ">=3.12"
dependencies = [
    "typer>=0.15.0",
    "pydantic>=2.10.0",
    "python-dateutil>=2.9.0.post0",
    "boto3>=1.35.0",
    "spacy>=3.8.0",
]

[project.scripts]
boston-pipeline = "pipeline.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pipeline"]

[tool.ruff]
target-version = "py312"
line-length = 120
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]

[tool.ruff.lint.isort]
known-first-party = ["pipeline"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = ["dateutil.*"]
ignore_missing_imports = true
disallow_untyped_calls = false

[tool.pytest.ini_options]
testpaths = ["tests"]

[dependency-groups]
dev = [
    "ruff>=0.9.0",
    "mypy>=1.14.0",
    "pytest>=8.3.0",
    "moto[s3]>=5.0.0",
    "boto3-stubs[s3]>=1.35.0",
]
